requirements.txt
Django==1.5.1
MarkupSafe==0.18
PIL==1.1.7
Pillow==2.1.0
colorama==0.2.7
daemonize==2.2.1
django-registration==1.0
django-simple-captcha==0.3.8
django-wsgiserver==0.8.0rc1
lxml==3.2.1
psutil==1.2.1
pyuv==0.10.8
requests==1.2.3
setproctitle==1.1.8
six==1.3.0
wsgiref==0.1.2